# pwnedapi/static/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PwnedAPI",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://api.pwnedhub.com",
      "variables": {}
    }
  ],
  "paths": {
    "/access-token": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Token Create (Login)",
        "operationId": "TokenCreate(Login)",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenCreateLoginRequest"
              },
              "example": {
                "username": "",
                "password": ""
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Misc"
        ],
        "summary": "Token Delete (Logout)",
        "operationId": "TokenDelete(Logout)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Users Read",
        "operationId": "UsersRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "User Create (Register)",
        "operationId": "UserCreate(Register)",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateRegisterRequest"
              },
              "example": {
                "username": "tim",
                "email": "tjt1980@gmail.com",
                "name": "Tim",
                "avatar": "https://www.gravatar.com/avatar/0a6d9b1ad59ad436bf9d9d16b2a7133e.png?s=300",
                "signature": "This is my signature.",
                "password": "password"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
    "/password-reset": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Password Reset Create",
        "operationId": "PasswordResetCreate",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "http://test.pwnedhub.com"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetCreateRequest"
              },
              "example": {
                "credential": "tim"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/notes": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Note Read",
        "operationId": "NoteRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Misc"
        ],
        "summary": "Note Update",
        "operationId": "NoteUpdate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteUpdateRequest"
              },
              "example": {
                "content": "##### Welcome to PwnedHub 2.0!\n\nA collaborative space to conduct hosted security assessments.\n\n**Find flaws.**\n\n* This is your notes space. Keep your personal notes here.\n* Leverage popular security testing tools right from your browser in the tools space.\n\n**Collaborate.**\n\n* Privately collborate with coworkers in the messaging space.\n* Join public rooms in the messaging space to share information and socialize.\n\n**On the Move**\n\n* PwnedHub 2.0 is built with mobility in mind. No need for a separate app!\n\nHappy hunting!\n\n\\- The PwnedHub Team\n\n"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/rooms": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Rooms Read",
        "operationId": "RoomsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Room Create",
        "operationId": "RoomCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreateRequest"
              },
              "example": {
                "name": "1:5",
                "private": true,
                "members": [
                  1,
                  5
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/unfurl": {
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Unfurl Create",
        "operationId": "UnfurlCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnfurlCreateRequest"
              },
              "example": {
                "url": "https://www.slack.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/tools": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Tools Read",
        "operationId": "ToolsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Tool Create",
        "operationId": "ToolCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCreateRequest"
              },
              "example": {
                "name": "ID",
                "path": "id",
                "description": "Displays the user ID of the current shell user."
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/scans": {
      "get": {
        "tags": [
          "Misc"
        ],
        "summary": "Scans Read",
        "operationId": "ScansRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Misc"
        ],
        "summary": "Scan Create",
        "operationId": "ScanCreate",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreateRequest"
              },
              "example": {
                "tid": 1,
                "args": "www.practisec.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "TokenCreateLoginRequest": {
        "title": "TokenCreateLoginRequest",
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "",
          "password": ""
        }
      },
      "UserCreateRegisterRequest": {
        "title": "UserCreateRegisterRequest",
        "required": [
          "username",
          "email",
          "name",
          "avatar",
          "signature",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          },
          "signature": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "username": "tim",
          "email": "tjt1980@gmail.com",
          "name": "Tim",
          "avatar": "https://www.gravatar.com/avatar/0a6d9b1ad59ad436bf9d9d16b2a7133e.png?s=300",
          "signature": "This is my signature.",
          "password": "password"
        }
      },
...
      "PasswordResetCreateRequest": {
        "title": "PasswordResetCreateRequest",
        "required": [
          "credential"
        ],
        "type": "object",
        "properties": {
          "credential": {
            "type": "string"
          }
        },
        "example": {
          "credential": "tim"
        }
      },
...
      "NoteUpdateRequest": {
        "title": "NoteUpdateRequest",
        "required": [
          "content"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "example": {
          "content": "##### Welcome to PwnedHub 2.0!\n\nA collaborative space to conduct hosted security assessments.\n\n**Find flaws.**\n\n* This is your notes space. Keep your personal notes here.\n* Leverage popular security testing tools right from your browser in the tools space.\n\n**Collaborate.**\n\n* Privately collborate with coworkers in the messaging space.\n* Join public rooms in the messaging space to share information and socialize.\n\n**On the Move**\n\n* PwnedHub 2.0 is built with mobility in mind. No need for a separate app!\n\nHappy hunting!\n\n\\- The PwnedHub Team\n\n"
        }
      },
      "RoomCreateRequest": {
        "title": "RoomCreateRequest",
        "required": [
          "name",
          "private",
          "members"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "members": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": ""
          }
        },
        "example": {
          "name": "1:5",
          "private": true,
          "members": [
            1,
            5
          ]
        }
      },
      "UnfurlCreateRequest": {
        "title": "UnfurlCreateRequest",
        "required": [
          "url"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          }
        },
        "example": {
          "url": "https://www.slack.com"
        }
      },
      "ToolCreateRequest": {
        "title": "ToolCreateRequest",
        "required": [
          "name",
          "path",
          "description"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "example": {
          "name": "ID",
          "path": "id",
          "description": "Displays the user ID of the current shell user."
        }
      },
      "ScanCreateRequest": {
        "title": "ScanCreateRequest",
        "required": [
          "tid",
          "args"
        ],
        "type": "object",
        "properties": {
          "tid": {
            "type": "integer",
            "format": "int32"
          },
          "args": {
            "type": "string"
          }
        },
        "example": {
          "tid": 1,
          "args": "www.practisec.com"
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}